Create OnePavementSensor_mach4.ino
Changed to the output to either print 'no activity' if frequency under 1.5hz (this is just a test frequency for now based on me shaking the sensor) or 'car detected' with the frequency attached. The differing frequencies will give data on the size of the cars and if the pavement is damaged.